Clean up implicits for simple imports
